// src/app/model/service-and-examine-model.ts
/**
 * Created by neng on 2/11/2559.
 */
export class ServiceAndExamineStatistic {

    agreementId: number;
    trNo: number;
    returnReqId: number;
    reqDate: Date|string;
    contractName:string ;
    createDate: Date|string;
    bookDate: Date|string;
    invokedDate: Date|string;
    returnedDate: Date|string;
    organize: string;
}